moving PluginBase into dedicated module to avoid circular imports
